personal-crud-ms/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
SQLAlchemy==2.0.22
pydantic==1.10.9
python-dotenv==1.0.0
psycopg2-binary==2.9.10
